refactor(api): move CertificationCandidateNotFoundError to certification/shared
