+++ api/tests/certification/enrolment/integration/infrastructure/repositories/candidate-repository_test.js
@@ -1,7 +1,7 @@
-import { CertificationCandidateNotFoundError } from '../../../../../../src/certification/enrolment/domain/errors.js';
 import { Candidate } from '../../../../../../src/certification/enrolment/domain/models/Candidate.js';
 import * as candidateRepository from '../../../../../../src/certification/enrolment/infrastructure/repositories/candidate-repository.js';
 import { SUBSCRIPTION_TYPES } from '../../../../../../src/certification/shared/domain/constants.js';
+import { CertificationCandidateNotFoundError } from '../../../../../../src/certification/shared/domain/errors.js';
 import { Frameworks } from '../../../../../../src/certification/shared/domain/models/Frameworks.js';
 import { _ } from '../../../../../../src/shared/infrastructure/utils/lodash-utils.js';
 import { expect } from '../../../../../test-helper.js';
