+++ api/tests/certification/enrolment/unit/domain/usecases/update-enrolled-candidate_test.js
@@ -1,11 +1,9 @@
 import sinon from 'sinon';
 
 import { updateEnrolledCandidate } from '../../../../../../src/certification/enrolment/domain/usecases/update-enrolled-candidate.js';
+import { CertificationCandidateNotFoundError } from '../../../../../../src/certification/shared/domain/errors.js';
 import { DomainTransaction } from '../../../../../../src/shared/domain/DomainTransaction.js';
-import {
-  CandidateAlreadyLinkedToUserError,
-  CertificationCandidateNotFoundError,
-} from '../../../../../../src/shared/domain/errors.js';
+import { CandidateAlreadyLinkedToUserError } from '../../../../../../src/shared/domain/errors.js';
 import { expect } from '../../../../../test-helper.js';
 import { domainBuilder } from '../../../../../tooling/domain-builder/domain-builder.js';
 import { catchErr } from '../../../../../tooling/test-utils/error.js';
