+++ api/tests/certification/session-management/integration/infrastructure/repositories/certification-candidate-for-supervising-repository_test.js
@@ -1,5 +1,5 @@
 import * as certificationCandidateForSupervisingRepository from '../../../../../../src/certification/session-management/infrastructure/repositories/certification-candidate-for-supervising-repository.js';
-import { CertificationCandidateNotFoundError } from '../../../../../../src/shared/domain/errors.js';
+import { CertificationCandidateNotFoundError } from '../../../../../../src/certification/shared/domain/errors.js';
 import { Assessment } from '../../../../../../src/shared/domain/models/Assessment.js';
 import { expect } from '../../../../../test-helper.js';
 import { databaseBuilder, knex } from '../../../../../tooling/databases.js';
