+++ api/tests/certification/shared/unit/application/http-error-mapper-configuration_test.js
@@ -1,6 +1,7 @@
 import { certificationDomainErrorMappingConfiguration } from '../../../../../src/certification/shared/application/http-error-mapper-configuration.js';
 import {
   CenterHabilitationError,
+  CertificationCandidateNotFoundError,
   CertificationCourseUpdateError,
   InvalidCertificationReportForFinalization,
 } from '../../../../../src/certification/shared/domain/errors.js';
@@ -58,6 +59,23 @@ describe('Unit | Certification | Shared | Application | HttpErrorMapperConfigura
         'This certification center has no habilitation for the given complementary certification.',
       );
       expect(error.code).to.equal('CENTER_HABILITATION_ERROR');
+    });
+  });
+
+  context('when mapping "CertificationCandidateNotFoundError"', function () {
+    it('returns a NotFoundError Http Error', function () {
+      //given
+      const httpErrorMapper = certificationDomainErrorMappingConfiguration.find(
+        (httpErrorMapper) => httpErrorMapper.name === CertificationCandidateNotFoundError.name,
+      );
+
+      //when
+      const error = httpErrorMapper.httpErrorFn(new CertificationCandidateNotFoundError());
+
+      //then
+      expect(error).to.be.instanceOf(HttpErrors.NotFoundError);
+      expect(error.message).to.equal('No candidate found');
+      expect(error.code).to.equal('CANDIDATE_NOT_FOUND');
     });
   });
 });
